update setup_symlinks script

+++ setup_symlinks.sh
@@ -1,7 +1,5 @@
 #!/bin/bash
 
-### TBH I haven't run this yet...b/c it's a little scary... I think it'll work tho...
-
 #####################
 #
 # This script copies all tracked dotfiles into target directory and creates symlinks
@@ -30,16 +28,23 @@ for file in "${files[@]}"; do
   # parse the directory name out of tracked file
   dir_name=`echo $file | sed -e 's_/[^/]*$__'`
 
-  # ensure nested directory structure exists in backup directory
+  # ensure nested directory structure exists in home and backup directories
   # (if file at source dir level, no need to create directory)
-  [ $dir_name != $file ] && mkdir -p "$backup_dir/$dir_name"
+  if [ $dir_name != $file ]; then
+    mkdir -p "$backup_dir/$dir_name"
+    mkdir -p "$HOME/$dir_name"
+  fi
 
-  if [ -f $target_dir/$file ]; then
-    echo "🤖 $file exists, backin'.it.up."
+  if [[ -f $target_dir/$file && ! -L $target_dir/$file ]]; then
+    echo "🤖 $file exists and isn't a symlink. Backin'.it.up."
     mv $target_dir/$file $backup_dir/
   fi
 
-  # TODO should ensure that symlink doesn't already exist, and destroy it if it does??
+  if [ -L $target_dir/$file ]; then
+    echo "symlink exists, removing $file 💣"
+    rm $target_dir/$file
+  fi
+
   echo "create symlink to git-tracked file 🔥"
   ln -s "$(pwd)/$file" $target_dir/$file
 done
